DPO-219 kamerakuvien käsittelyn parannuksia - fixed return type

+++ src/main/java/fi/livi/digitraffic/tie/conf/jms/CameraJMSListenerConfiguration.java
@@ -61,7 +61,7 @@ public class CameraJMSListenerConfiguration extends AbstractJMSListenerConfigura
         JMSMessageListener.JMSDataUpdater<Kuva> handleData = (List<Pair<Kuva, String>> data) -> {
             try {
                 List<Kuva> kuvaData = data.stream().map(Pair::getLeft).collect(Collectors.toList());
-                return (int) cameraDataUpdateService.updateCameraData(kuvaData);
+                return cameraDataUpdateService.updateCameraData(kuvaData);
             } catch (SQLException e) {
                 log.error("Error while handling Camera data", e);
                 return 0;
